chore: refactor!: migrate exceptions to `com.lsadf.core.exception` package
- Moved all exception classes from `com.lsadf.core.infra.exception` to `com.lsadf.core.exception` for improved module structure.
- Updated all related imports across the codebase to reflect the new package structure.
- Adjusted `package-info.java` files to match the updated package layout.
- Ensured functional consistency by modifying tests and configuration files accordingly.
