+++ lsadf_admin/src/test/java/com/lsadf/admin/application/bdd/step_definition/given/BddGivenStepDefinitions.java
@@ -24,7 +24,7 @@ import com.lsadf.core.bdd.BddUtils;
 import com.lsadf.core.domain.game.save.characteristics.Characteristics;
 import com.lsadf.core.domain.game.save.currency.Currency;
 import com.lsadf.core.domain.game.save.stage.Stage;
-import com.lsadf.core.infra.exception.http.NotFoundException;
+import com.lsadf.core.exception.http.NotFoundException;
 import com.lsadf.core.infra.persistence.impl.game.inventory.ItemEntity;
 import com.lsadf.core.infra.persistence.impl.game.save.characteristics.CharacteristicsEntity;
 import com.lsadf.core.infra.persistence.impl.game.save.currency.CurrencyEntity;
