+++ lsadf_core/src/main/java/com/lsadf/core/application/game/inventory/impl/InventoryServiceImpl.java
@@ -21,9 +21,9 @@ import com.lsadf.core.application.game.save.metadata.GameMetadataService;
 import com.lsadf.core.domain.game.inventory.item.Item;
 import com.lsadf.core.domain.game.inventory.item.ItemRarity;
 import com.lsadf.core.domain.game.inventory.item.ItemType;
-import com.lsadf.core.infra.exception.AlreadyExistingItemClientIdException;
-import com.lsadf.core.infra.exception.http.ForbiddenException;
-import com.lsadf.core.infra.exception.http.NotFoundException;
+import com.lsadf.core.exception.AlreadyExistingItemClientIdException;
+import com.lsadf.core.exception.http.ForbiddenException;
+import com.lsadf.core.exception.http.NotFoundException;
 import com.lsadf.core.infra.web.request.game.inventory.ItemRequest;
 import java.util.*;
 import org.springframework.transaction.annotation.Transactional;
