+++ lsadf_core/src/main/java/com/lsadf/core/application/game/save/characteristics/impl/CharacteristicsServiceImpl.java
@@ -20,7 +20,7 @@ import com.lsadf.core.application.game.save.characteristics.CharacteristicsRepos
 import com.lsadf.core.application.game.save.characteristics.CharacteristicsService;
 import com.lsadf.core.application.shared.CachePort;
 import com.lsadf.core.domain.game.save.characteristics.Characteristics;
-import com.lsadf.core.infra.exception.http.NotFoundException;
+import com.lsadf.core.exception.http.NotFoundException;
 import java.util.Optional;
 import java.util.UUID;
 import org.springframework.transaction.annotation.Transactional;
