+++ lsadf_core/src/main/java/com/lsadf/core/application/game/save/currency/CurrencyService.java
@@ -16,7 +16,7 @@
 package com.lsadf.core.application.game.save.currency;
 
 import com.lsadf.core.domain.game.save.currency.Currency;
-import com.lsadf.core.infra.exception.http.NotFoundException;
+import com.lsadf.core.exception.http.NotFoundException;
 import java.util.UUID;
 
 public interface CurrencyService {
