+++ lsadf_core/src/main/java/com/lsadf/core/application/game/save/impl/GameSaveServiceImpl.java
@@ -32,11 +32,11 @@ import com.lsadf.core.domain.game.save.characteristics.Characteristics;
 import com.lsadf.core.domain.game.save.currency.Currency;
 import com.lsadf.core.domain.game.save.metadata.GameMetadata;
 import com.lsadf.core.domain.game.save.stage.Stage;
-import com.lsadf.core.infra.exception.AlreadyExistingGameSaveException;
-import com.lsadf.core.infra.exception.AlreadyTakenNicknameException;
-import com.lsadf.core.infra.exception.http.ForbiddenException;
-import com.lsadf.core.infra.exception.http.NotFoundException;
-import com.lsadf.core.infra.exception.http.UnauthorizedException;
+import com.lsadf.core.exception.AlreadyExistingGameSaveException;
+import com.lsadf.core.exception.AlreadyTakenNicknameException;
+import com.lsadf.core.exception.http.ForbiddenException;
+import com.lsadf.core.exception.http.NotFoundException;
+import com.lsadf.core.exception.http.UnauthorizedException;
 import com.lsadf.core.infra.web.request.game.characteristics.CharacteristicsRequest;
 import com.lsadf.core.infra.web.request.game.currency.CurrencyRequest;
 import com.lsadf.core.infra.web.request.game.metadata.GameMetadataRequest;
