+++ lsadf_core/src/main/java/com/lsadf/core/exception/AlreadyExistingGameSaveException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lsadf.core.infra.exception;
+package com.lsadf.core.exception;
 
 /** Exception for when a game save already exists */
 public class AlreadyExistingGameSaveException extends RuntimeException {
