+++ lsadf_core/src/main/java/com/lsadf/core/exception/AlreadyTakenNicknameException.java
@@ -13,19 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lsadf.core.infra.exception.http;
+package com.lsadf.core.exception;
 
-/** Exception for when a resource is not found. */
-public class NotFoundException extends RuntimeException {
-  public NotFoundException() {
-    super();
-  }
-
-  public NotFoundException(String message) {
+/** Exception for when a nickname is already taken */
+public class AlreadyTakenNicknameException extends RuntimeException {
+  public AlreadyTakenNicknameException(String message) {
     super(message);
-  }
-
-  public NotFoundException(String message, Throwable cause) {
-    super(message, cause);
   }
 }
