+++ lsadf_core/src/main/java/com/lsadf/core/exception/http/UnauthorizedException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lsadf.core.infra.exception.http;
+package com.lsadf.core.exception.http;
 
 /** Authentication Exception */
 public class UnauthorizedException extends RuntimeException {
