+++ lsadf_core/src/main/java/com/lsadf/core/infra/web/controller/advice/GlobalExceptionHandler.java
@@ -17,8 +17,11 @@ package com.lsadf.core.infra.web.controller.advice;
 
 import static com.lsadf.core.infra.web.response.ResponseUtils.generateResponse;
 
-import com.lsadf.core.infra.exception.*;
-import com.lsadf.core.infra.exception.http.*;
+import com.lsadf.core.exception.AlreadyExistingGameSaveException;
+import com.lsadf.core.exception.AlreadyExistingItemClientIdException;
+import com.lsadf.core.exception.AlreadyExistingUserException;
+import com.lsadf.core.exception.AlreadyTakenNicknameException;
+import com.lsadf.core.exception.http.*;
 import com.lsadf.core.infra.web.response.ApiResponse;
 import java.util.HashMap;
 import java.util.Map;
