+++ lsadf_core/src/test/java/com/lsadf/core/unit/application/game/inventory/InventoryServiceTests.java
@@ -25,8 +25,8 @@ import com.lsadf.core.application.game.inventory.InventoryRepositoryPort;
 import com.lsadf.core.application.game.inventory.impl.InventoryServiceImpl;
 import com.lsadf.core.application.game.save.metadata.GameMetadataService;
 import com.lsadf.core.domain.game.inventory.item.*;
-import com.lsadf.core.infra.exception.AlreadyExistingItemClientIdException;
-import com.lsadf.core.infra.exception.http.NotFoundException;
+import com.lsadf.core.exception.AlreadyExistingItemClientIdException;
+import com.lsadf.core.exception.http.NotFoundException;
 import com.lsadf.core.infra.web.request.game.inventory.ItemRequest;
 import java.util.*;
 import org.junit.jupiter.api.BeforeEach;
