+++ lsadf_core/src/test/java/com/lsadf/core/unit/application/game/save/CurrencyServiceTests.java
@@ -28,7 +28,7 @@ import com.lsadf.core.application.game.save.currency.CurrencyService;
 import com.lsadf.core.application.game.save.currency.impl.CurrencyServiceImpl;
 import com.lsadf.core.application.shared.CachePort;
 import com.lsadf.core.domain.game.save.currency.Currency;
-import com.lsadf.core.infra.exception.http.NotFoundException;
+import com.lsadf.core.exception.http.NotFoundException;
 import java.util.Optional;
 import java.util.UUID;
 import org.jspecify.annotations.NonNull;
